Header file moved in opm-output.
